fix up file paths for features folder and added floating action button
- fixed up path in features by adding community, home, and learn folder
and put respective files in folder
- refactored widgets into widget folder in features folder
- added floating action button for community to add a post
