+++ lib/router/routes.dart
@@ -3,7 +3,7 @@ import 'package:go_router/go_router.dart';
 import '../views/auth/forgot_password_screen.dart';
 import '../views/auth/login_screen.dart';
 import '../views/auth/signup_screen.dart';
-import '../views/features/home/widgets/home_scaffold.dart';
+import '../views/features/widgets/home_scaffold.dart';
 import '../views/landing_screen.dart';
 
 class AppRouter {
