+++ lib/views/features/widgets/tab_bar_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:recycle/views/features/home/community_screen.dart';
-import 'package:recycle/views/features/home/learn_screen.dart';
+import 'package:recycle/views/features/community/community_screen.dart';
+import 'package:recycle/views/features/learn/learn_screen.dart';
 
 // utils
 import 'package:recycle/utils/providers/login_state_provider.dart';
